refactor: konum değişikliği ve gereksiz kodlar temizlendi
