+++ src/components/index.ts
@@ -1,5 +1,5 @@
-export { pointInfoEmbed } from './pointInfoEmbed';
-export { userPointsEmbed } from './userPointsEmbed';
+export { pointInfoEmbed } from './embeds/point/pointInfoEmbed';
+export { userPointsEmbed } from './embeds/point/userPointsEmbed';
 
 export * from './embeds';
 export * from './buttons';
