+++ src/slash-commands/index.ts
@@ -1,5 +1,5 @@
 export { PointInfo } from './pointInfo/PointInfo';
 export { UserPoints } from './userPoints/UserPoints';
-export { SlashPing } from './SlashPing/SlashPing';
+
 export { Setup } from './setup/Setup';
 export { Settings } from './settings/Settings';
